refactor(ZNTA-1905): update relative imports for moved tests and modules
Correct all relative imports to match new file locations. Kept as a
separate commit to ensure these changes do not interfere with git's
detection of renamed files.
