+++ server/zanata-frontend/src/frontend/app/editor/components/IconButtonToggle/index.js
@@ -20,7 +20,7 @@
  */
 
 import cx from 'classnames'
-import IconButton from './IconButton'
+import IconButton from '../IconButton'
 import React, { PropTypes } from 'react'
 
 /**
