+++ server/zanata-frontend/src/frontend/app/editor/components/Pager/Pager.test.js
@@ -3,9 +3,9 @@ jest.disableAutomock()
 import React from 'react'
 import ReactDOMServer from 'react-dom/server'
 import TestUtils from 'react-addons-test-utils'
-import Pager from '../../app/editor/components/Pager'
+import Pager from '.'
 import { Icon } from 'zanata-ui'
-import mockGettextCatalog from '../mock/mockAngularGettext'
+import mockGettextCatalog from '../../../../__tests__/mock/mockAngularGettext'
 
 describe('PagerTest', () => {
   it('Pager markup', () => {
