+++ server/zanata-frontend/src/frontend/app/utils/GlossaryHelper.test.js
@@ -1,6 +1,6 @@
 jest.disableAutomock()
 
-import GlossaryHelper from '../../app/utils/GlossaryHelper'
+import GlossaryHelper from './GlossaryHelper'
 
 describe('GlossaryHelperTest', function () {
   // var GlossaryHelper
